Chapter 4
Just learned through this exercise a subtle difference in how Ruby handles
*protected* / *private* methods compared to other languages (such as Java).

See Classes, Objects, and Variables - Access Control at http://ruby-doc.com/docs/ProgrammingRuby/

> "If a method is private, it may be called only within the context of the calling
object---it is never possible to access another object's private methods directly,
even if the object is of the same class as the caller.

This is why the `attr_reader :amount`  is set to *protected*, so that
`other.amount` can be called from `Dollar#==`.